+++ app/Http/Controllers/PaymentController.php
@@ -25,11 +25,11 @@ class PaymentController extends Controller
         $data = DB::table('Payment')
             ->paginate(10);
         //->get();
-        
+
         $user = Auth::user();
         $check = $this->checkAccess('payment.index', $user->id, $user->idRole);
         if ($check) {
-            return view('master.payment.index',[
+            return view('master.payment.index', [
                 'data' => $data,
             ]);
         } else {
@@ -45,7 +45,7 @@ class PaymentController extends Controller
     public function create()
     {
         //
-        
+
         $user = Auth::user();
         $check = $this->checkAccess('payment.create', $user->id, $user->idRole);
         if ($check) {
@@ -66,18 +66,19 @@ class PaymentController extends Controller
         //
         $data = $request->collect();
         $user = Auth::user();
-        
+
         DB::table('Payment')
-            ->insert(array(
-                'Name' => $data['name'],
-                'Deskripsi' => $data['deskripsi'],
-                'CreatedBy'=> $user->id,
-                'CreatedOn'=> date("Y-m-d h:i:sa"),
-                'UpdatedBy'=> $user->id,
-                'UpdatedOn'=> date("Y-m-d h:i:sa"),
-            )
-        );
-        return redirect()->route('payment.index')->with('status','Success!!');
+            ->insert(
+                array(
+                    'Name' => $data['name'],
+                    'Deskripsi' => $data['deskripsi'],
+                    'CreatedBy' => $user->id,
+                    'CreatedOn' => date("Y-m-d h:i:sa"),
+                    'UpdatedBy' => $user->id,
+                    'UpdatedOn' => date("Y-m-d h:i:sa"),
+                )
+            );
+        return redirect()->route('payment.index')->with('status', 'Success!!');
     }
 
     /**
@@ -89,11 +90,11 @@ class PaymentController extends Controller
     public function show(Payment $payment)
     {
         //
-        
+
         $user = Auth::user();
         $check = $this->checkAccess('payment.show', $user->id, $user->idRole);
         if ($check) {
-            return view('master.payment.detail',[
+            return view('master.payment.detail', [
                 'payment' => $payment,
             ]);
         } else {
@@ -110,12 +111,12 @@ class PaymentController extends Controller
     public function edit(Payment $payment)
     {
         //
-        
+
 
         $user = Auth::user();
         $check = $this->checkAccess('payment.edit', $user->id, $user->idRole);
         if ($check) {
-            return view('master.payment.edit',[
+            return view('master.payment.edit', [
                 'payment' => $payment,
             ]);
         } else {
@@ -137,14 +138,15 @@ class PaymentController extends Controller
         $user = Auth::user();
         DB::table('Payment')
             ->where('PaymentID', $payment['PaymentID'])
-            ->update(array(
-                'Name' => $data['name'],
-                'Deskripsi' => $data['deskripsi'],
-                'UpdatedBy'=> $user->id,
-                'UpdatedOn'=> date("Y-m-d h:i:sa"),
+            ->update(
+                array(
+                    'Name' => $data['name'],
+                    'Deskripsi' => $data['deskripsi'],
+                    'UpdatedBy' => $user->id,
+                    'UpdatedOn' => date("Y-m-d h:i:sa"),
                 )
-        );  
-        return redirect()->route('payment.index')->with('status','Success!!');
+            );
+        return redirect()->route('payment.index')->with('status', 'Success!!');
     }
 
     /**
@@ -156,22 +158,28 @@ class PaymentController extends Controller
     public function destroy(Payment $payment)
     {
         //
-        $payment->delete();
-        return redirect()->route('payment.index')->with('status','Success!!');
+        $user = Auth::user();
+        $check = $this->checkAccess('payment.edit', $user->id, $user->idRole);
+        if ($check) {
+            $payment->delete();
+            return redirect()->route('payment.index')->with('status', 'Success!!');
+        } else {
+            return redirect()->route('home')->with('message', 'Anda tidak memiliki akses kedalam Jenis Pembayaran');
+        }
     }
 
     public function searchPaymentName(Request $request)
     {
         $name = $request->input('searchname');
         $data = DB::table('Payment')
-            ->where('Name','like','%'.$name.'%')
+            ->where('Name', 'like', '%' . $name . '%')
             ->paginate(10);
         //->get();
-        
+
         $user = Auth::user();
         $check = $this->checkAccess('payment.index', $user->id, $user->idRole);
         if ($check) {
-            return view('master.payment.index',[
+            return view('master.payment.index', [
                 'data' => $data,
             ]);
         } else {
